requirements.txt
django
djangorestframework
djangorestframework-simplejwt
Pillow
django-simple-history